adding SongsCards component
