+++ scripts/components/Songs.js
@@ -1,22 +1,14 @@
 import React, {Component, PropTypes} from 'react';
 import {connect} from 'react-redux';
 
-import {playSong} from '../actions/player';
 import {fetchSongsIfNeeded} from '../actions/playlists';
 
 import InfiniteScrollify from '../components/InfiniteScrollify';
-import SongsCard from '../components/SongsCard';
-import Spinner from '../components/Spinner';
+import SongsCards from '../components/SongsCards';
 import Stickify from '../components/Stickify';
 import Toolbar from '../components/Toolbar';
 
-
 class Songs extends Component {
-    constructor(props) {
-        super(props);
-        this.renderSongs = this.renderSongs.bind(this);
-    }
-
     componentWillMount() {
         const {dispatch, playlist} = this.props;
         dispatch(fetchSongsIfNeeded(playlist));
@@ -31,60 +23,20 @@ class Songs extends Component {
         }
     }
 
-    playSong(i) {
-        const {playlist, dispatch} = this.props;
-        dispatch(playSong(playlist, i));
-    }
-
-    renderSongs() {
-        const chunk = 5;
-        const {dispatch, playlist, playlists, playingSongId, songs, users} = this.props;
-        const items = playlist in playlists ? playlists[playlist].items : [];
-
-        let result = [];
-        for (let i = 0; i < items.length; i += chunk) {
-            let songCards = items.slice(i, i + chunk).map((songId, j) => {
-                const song = songs[songId];
-                const user = users[song.user_id];
-                const index = i + j;
-                return (
-                    <div className='col-1-5' key={index + '-' + song.id}>
-                        <SongsCard
-                            dispatch={dispatch}
-                            isActive={song.id === playingSongId}
-                            playSong={this.playSong.bind(this, index)}
-                            song={song}
-                            user={user} />
-                    </div>
-                );
-            }, this);
-
-            if (songCards.length < chunk) {
-                for (let j = 0; j < chunk - songCards.length + 1; j++) {
-                    songCards.push(<div className='col-1-5' key={'song-placeholder-' + (i + j)}></div>);
-                }
-            }
-
-            result.push(
-                <div className='songs-row grid' key={'songs-row-' + i}>{songCards}</div>
-            );
-        }
-
-        return result;
-    }
-
     render() {
-        const {dispatch, playlist, playlists, sticky, time} = this.props;
-        const isFetching = playlist in playlists ? playlists[playlist].isFetching : false;
+        const {dispatch, playingSongId, playlist, playlists, sticky, songs, time, users} = this.props;
 
         return (
             <div className={'songs' + (sticky ? ' sticky' : '')}>
                 <Toolbar dispatch={dispatch} playlist={playlist} sticky={sticky} time={time} />
                 <div className='container'>
-                    <div className='content'>
-                        {this.renderSongs()}
-                        {isFetching ? <Spinner /> : null}
-                    </div>
+                    <SongsCards
+                        dispatch={dispatch}
+                        playingSongId={playingSongId}
+                        playlist={playlist}
+                        playlists={playlists}
+                        songs={songs}
+                        users={users} />
                 </div>
             </div>
         );
